rename functional te executor and move transform
